some fixes

+++ 20290229-WEB-API/Migrations/20231116202001_Orders.cs
@@ -39,7 +39,7 @@ namespace VerifyEmailForgotPasswordTutorial.Migrations
                     OrderNote = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     OrderCreatingTime = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     ItemCount = table.Column<int>(type: "int", nullable: false),
-                    OrderTotalPrice = table.Column<double>(type: "float", nullable: false)
+                    OrderTotalPrice = table.Column<double>(type: "double", nullable: false)
                 },
                 constraints: table =>
                 {
